Make Input and Event/EventStream internal
This moves the Input, Event and EventStream classes
into a new internal header.

To decouple this I had to describe the type of
the priority queue in TurboEvents in a different
way which gives some domino effects in the other
parts of the code, but there should be no functional
change.

External users of the library should continue to include
turboevents.hpp, things inside lib/ should include
turboevents-internal.hpp instead after this change.
