+++ lib/IO/CMakeLists.txt
@@ -1,3 +1,5 @@
+target_include_directories(turboevents PRIVATE ${TurboEvents_SOURCE_DIR}/lib)
+
 find_package(PkgConfig REQUIRED)
 pkg_check_modules(kafka REQUIRED IMPORTED_TARGET rdkafka)
 target_sources(turboevents PRIVATE KafkaOutput.cpp)
