+++ lib/IO/KafkaOutput.hpp
@@ -1,4 +1,4 @@
-#include "turboevents.hpp"
+#include "turboevents-internal.hpp"
 
 namespace TurboEvents {
